No need for this to be a double
Working on eliminating compiler warnings for FP comparisons, this variable didn't need to be a double anyway.

+++ SquareDesk-DEV/test123/mainwindow_sd.cpp
@@ -225,7 +225,7 @@ static void decode_formation_into_dancer_destinations(const QStringList &sdforma
     {
         int dancer_start_x = -1;
         bool draw = false;
-        double direction = -1.0;
+        int direction = -1.0;
 
         for (int x = 0; x < sdformation[y].length(); ++x)
         {
